+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Track clientId for tracing purposes [#261](https://github.com/CartoDB/carto-react/pull/261)
 ## 1.2.0-alpha.0 (2021-12-29)
 - WRAPPER WIDGET Configurable margin in widgets [#251](https://github.com/CartoDB/carto-react/pull/251)
 - Add spatial filter to allow ramdom geometry filters in widgets[#250](https://github.com/CartoDB/carto-react/pull/250)